requirements.txt
click==8.1.7
pyyaml==6.0.2
gitpython==3.1.43
tree-sitter==0.20.4
tree_sitter_languages==1.10.2